Move mpipool_support.py into autoparallelize
